zakladni verze magery hotova

+++ moriaengine/scripts/systems/skills/Healing/Healing.cs
@@ -170,12 +170,12 @@ namespace SteamEngine.CompiledScripts {
 				return;
 			}
 
-			ushort selfHealing = self.GetSkill((int) SkillName.Healing);
+			ushort selfHealing = self.GetSkill(SkillName.Healing);
 
 			//pokud se leceni zdarilo
 			if (SingletonScript<HealingSkillDef>.Instance.CheckSuccess(self, Globals.dice.Next(200))) {
 
-				targetted.Hits += (short) ScriptUtil.EvalRangePermille(self.GetSkill((int) SkillName.Healing) + self.GetSkill((int) SkillName.Anatomy), SingletonScript<HealingSkillDef>.Instance.Effect);
+				targetted.Hits += (short) ScriptUtil.EvalRangePermille(self.GetSkill(SkillName.Healing) + self.GetSkill(SkillName.Anatomy), SingletonScript<HealingSkillDef>.Instance.Effect);
 
 				if (targetted.Hits > targetted.MaxHits)
 					targetted.Hits = targetted.MaxHits;
